+++ src/main/java/com/project/todayWhatToDo/user/service/CareerService.java
@@ -1,7 +1,12 @@
 package com.project.todayWhatToDo.user.service;
 
+import com.project.todayWhatToDo.user.domain.Career;
+import com.project.todayWhatToDo.user.domain.Company;
+import com.project.todayWhatToDo.user.domain.User;
+import com.project.todayWhatToDo.user.dto.CreateCareerRequestDto;
 import com.project.todayWhatToDo.user.dto.UpdateCareerRequestDto;
 import com.project.todayWhatToDo.user.exception.CareerNotFoundException;
+import com.project.todayWhatToDo.user.exception.UserNotFoundException;
 import com.project.todayWhatToDo.user.repository.CareerRepository;
 import com.project.todayWhatToDo.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -24,5 +29,25 @@ public class CareerService {
 
     public void deleteCareer(Long careerId, Long userId){
         careerRepository.deleteByIdAndUserId(careerId, userId);
+    }
+
+    public void createCareer(CreateCareerRequestDto request) {
+        User user = userRepository.findById(request.userId())
+                .orElseThrow(UserNotFoundException::new);
+
+        user.addCareer(
+                Career.builder()
+                        .user(user)
+                        .company(Company.builder()
+                                .name(request.name())
+                                .address(request.address())
+                                .build()
+                        )
+                        .introduction(request.introduction())
+                        .startedAt(request.startedAt())
+                        .endedAt(request.endedAt())
+                        .position(request.position())
+                        .build()
+        );
     }
 }
